requirements.txt
google-api-python-client
google-auth
edge-tts
